Some fixes and notebook for model uncertainty detectors
